Refactor: TodoItem 컴포넌트 분리 및 모바일 반응형 스타일 수정 (#216)
* #215 refactor: 멤버 프로필 이미지 gap 제거

* #215 refactor: 투두 아이템 컴포넌트 분리 및 모바일 스타일 개선

- TodoItem 컴포넌트를 분리하여 가독성 및 재사용성 향상
- 모바일 환경 기준 폰트 사이즈 및 width 조정
